requirements.txt
gradio
Pillow
rembg
numpy